Purify fix
Added fix for stubborn purity mode which was setting itself on regardless of picked
mode. Crucial was the last byte (right before CRC) which was set to 0x0C.
In case of Haier Flexis White Matt (firmware R_1.0.00/e_2.5.14) it is
always set to zero (check with UART2USB dongle). Also set CRC bytes to
zero and status data byte, which always are overriden.