Hack 22 degrees case
For some reason in captured data when crc16 is computed to FF 29
there is a 55 in the middle -> FF 55 29. As a "temporary" workaround just
hardcore this case. This is ugly hack, but I don't have any idea what
is this number. It happens when temperature is set to 22 and fans to
auto.